rf function in command utils and moved them to appropriate files
