+++ products/utils.py
@@ -1,0 +1,18 @@
+import re
+from utils.helper_func import clean_data
+
+
+def get_product_price(item):
+    regex = r"[\-:;]{1}"
+    item = clean_data(item)
+    delimiter = re.findall(regex, item)[0]
+    if delimiter is None or delimiter == "":
+        return {
+            "success": False,
+            "message": "{}: please use a valid delimiter like - or , or : example. produce - amount".format(
+                item
+            ),
+        }
+    else:
+        product, amount = item.split(delimiter)
+        return {"success": True, "data": {"price_amount": [product, amount]}}
