+++ utils/helper_func.py
@@ -1,0 +1,2 @@
+def clean_data(data):
+    return data.strip().lower()
